rust: Fix issue with incorrect utf parsing
There was an issue when parser encountered `©` char in string, It is
represented as two bytes and we're copying bytes after byte when parsing
string, and copy each byte into `String` buffer. This probably doesn't
play well with Strings, which is utf-aware. The end result was that `©`
after copying byte after byte ended up as two separate characters `Â©`.